fix: np.ceil returns non-int object

+++ src/utils/data_utils.py
@@ -34,7 +34,7 @@ def split_classes_for_tasks(num_classes: int, num_tasks: int, num_classes_per_ta
         warnings.warn(f'We will have duplicated classes: {num_classes_to_use} > {num_classes - num_reserved_classes}')
 
     classes = np.arange(num_classes)[num_reserved_classes:]
-    classes = np.tile(classes, np.ceil(num_classes_to_use / len(classes)))[:num_classes_to_use]
+    classes = np.tile(classes, np.ceil(num_classes_to_use / len(classes)).astype(int))[:num_classes_to_use]
     splits = np.random.permutation(classes).reshape(num_tasks, num_classes_per_task)
 
     return splits
